move USER_INTENT enum to constants folder
