+++ app/components/Views/AccountConnect/AccountConnectMultiSelector/AccountConnectMultiSelector.tsx
@@ -21,7 +21,7 @@ import ButtonLink from '../../../../component-library/components/Buttons/Button/
 import styleSheet from './AccountConnectMultiSelector.styles';
 import { AccountConnectMultiSelectorProps } from './AccountConnectMultiSelector.types';
 import { ButtonSecondaryVariants } from '../../../../component-library/components/Buttons/Button/variants/ButtonSecondary';
-import { USER_INTENT } from '../AccountConnect';
+import USER_INTENT from '../../../../constants/permissions';
 
 const AccountConnectMultiSelector = ({
   accounts,
