+++ app/components/Views/AccountConnect/AccountConnectSingleSelector/AccountConnectSingleSelector.types.ts
@@ -1,6 +1,6 @@
 // External dependencies.
 import { UseAccounts } from '../../../hooks/useAccounts';
-import { USER_INTENT } from '../AccountConnect';
+import USER_INTENT from '../../../../constants/permissions';
 import { AccountConnectScreens } from '../AccountConnect.types';
 
 /**
