+++ app/constants/permissions.ts
@@ -1,0 +1,11 @@
+enum USER_INTENT {
+  None,
+  Create,
+  CreateMultiple,
+  Confirm,
+  Cancel,
+  Import,
+  ConnectHW,
+}
+
+export default USER_INTENT;
